Revert defaultValue support
After my reflection, this design is not a good idea.
This is because users may use u2x function as a callback of Array.prototype.map, that the second parameter indicates the index of current array item.
If the second parameter of u2x is designed as a default value, the array index will be passed incorrectly and no any errors are reported in the development phase.
For example, `[ 123, null, 456, null ].map(u2n)` will result `[ 123, 1, 456, 3 ]`, this is very dangerous.

A better solution is use TypeScript native operator `??` to set the default value.
For example, `u2n(what) ?? -1`, when `what` is not a number, `u2n` will return an `undefined`, and then the `??` operator causes it fallback to final `-1` as the default value.